clean up Collection query renderers
